+++ transports/websocket/CHANGELOG.md
@@ -48,5 +48,5 @@
 
 # 0.20.0 [2020-06-22]
 
-- Updated `soketto` dependency which caused some smaller
+- Updated `soket` dependency which caused some smaller
   API changes ([PR 1603](https://github.com/libp2p/rust-libp2p/pull/1603)).